Some clarifications on how to deploy group config
Being following the documentation from top to bottom, it is easy to be confused and think that to deploy an application configuration based on groups is the same as deploying a simple application. So better to clarify that now we are using the /v2/groups endpoint instead of the /v2/apps.